Refactor all response status

+++ actions/new-password.ts
@@ -8,36 +8,65 @@ import {
 import { getUserByEmail, updateUserById } from "@/services/user";
 import { redirect } from "next/navigation";
 import { z } from "zod";
-import { hashPassword, isExpired } from "@/lib/utils";
+import { hashPassword, isExpired, response } from "@/lib/utils";
 
 export const newPassword = async (payload: z.infer<typeof newPasswordSchema>, token: string) => {
+  // Check if user input is not valid, then return an error.
   const validatedFields = newPasswordSchema.safeParse(payload);
   if (!validatedFields.success) {
-    return { error: "Invalid fields." };
+    return response({
+      success: false,
+      error: {
+        code: 422,
+        message: "Invalid fields.",
+      },
+    });
   }
 
   const { password } = validatedFields.data;
 
+  // Check if token doesn't exist, then redirect to login page.
   const existingToken = await getResetPasswordToken(token);
   if (!existingToken) redirect("/");
 
+  // Check if token has expired, then return an error.
   const hasExpired = isExpired(existingToken.expires);
   if (hasExpired) {
-    return { error: "Token has expired. Please resend to your email." };
+    return response({
+      success: false,
+      error: {
+        code: 401,
+        message: "Token has expired. Please resend to your email.",
+      },
+    });
   }
 
+  // Check if email address doesn't exist, then return an error.
   const existingUser = await getUserByEmail(existingToken.email);
-  if (!existingUser) {
-    return { error: "Email does not exist." };
+  if (!existingUser || !existingUser.email || !existingUser.password) {
+    return response({
+      success: false,
+      error: {
+        code: 401,
+        message: "Email address does not exist.",
+      },
+    });
   }
 
+  // Create new password by hashing the password first.
   const hashedPassword = await hashPassword(password);
 
+  // Replace the old password with the new one.
   await updateUserById(existingUser.id, {
     password: hashedPassword,
   });
-
+  // Delete reset password token.
   await deleteResetPasswordTokenById(existingToken.id);
 
-  return { success: "Your password has been reset successfully." };
+  // Then return response success.
+  return response({
+    success: true,
+    code: 200,
+    message: "Your password has been reset successfully.",
+  });
 };
